refactor: merge Harbor health into harbor client package (#595)
- Move health check logic and types from internal/groundcontrol/harborhealth into internal/groundcontrol/harbor/health.go
- Remove internal/groundcontrol/harborhealth directory
- Update cmd/groundcontrol/server/main.go to call harbor.CheckHealth()
